refactor kde-related functions and small fixes
